fix: native session honors strict4k envelope prompt
Native mode was falling back to raw conversation history when structured
messages were nil, skipping ContextCore windowing and PromptEnvelope
truncation. Match capture mode and move the Apple session path off Agent.swift.
