+++ Sources/Swarm/Agents/Agent.swift
@@ -5,10 +5,6 @@
 
 import Foundation
 
-#if canImport(FoundationModels)
-import FoundationModels
-#endif
-
 // MARK: - Agent
 
 /// An agent that uses structured LLM tool calling APIs for reliable tool invocation.
@@ -633,13 +629,13 @@ public struct Agent: AgentRuntime, Sendable {
         let structuredOutput: StructuredOutputResult?
     }
 
-    private struct ToolLoopOutcome: Sendable {
+    struct ToolLoopOutcome: Sendable {
         let output: String
         let structuredOutput: StructuredOutputResult?
         let transcriptMessages: [MemoryMessage]
     }
 
-    private struct FinalAssistantResponse: Sendable {
+    struct FinalAssistantResponse: Sendable {
         let content: String
         let structuredOutput: StructuredOutputResult?
     }
@@ -1262,7 +1258,7 @@ public struct Agent: AgentRuntime, Sendable {
         )
     }
 
-    private func finalizeAssistantResponse(
+    func finalizeAssistantResponse(
         content: String,
         request: StructuredOutputRequest?,
         provider: any InferenceProvider
@@ -1488,7 +1484,10 @@ public struct Agent: AgentRuntime, Sendable {
 
                 if let nativeOutcome = try await executeNativeFoundationModelsSessionIfAvailable(
                     provider: provider,
-                    messages: structuredMessages ?? conversationHistory.map(\.inferenceMessage),
+                    messages: FoundationModelsNativePrompt.messages(
+                        structuredMessages: structuredMessages,
+                        envelopePrompt: prompt
+                    ),
                     toolRegistry: toolRegistry,
                     toolSchemas: toolSchemas,
                     inferenceOptions: inferenceOptions,
@@ -1662,148 +1661,6 @@ public struct Agent: AgentRuntime, Sendable {
         return history
     }
 
-    /// Runs Foundation Models native session mode when opted in and the provider
-    /// is ``FoundationModelsInferenceProvider``. Returns `nil` so the Swarm loop
-    /// continues when the flag is `.capture` or the provider is not FM.
-    ///
-    /// Memory has already been injected into `messages` (session start). Swarm's
-    /// iteration cap, mid-loop checkpoints, and per-turn guardrail wrapping are
-    /// not applied inside Apple's tool loop.
-    private func executeNativeFoundationModelsSessionIfAvailable(
-        provider: any InferenceProvider,
-        messages: [InferenceMessage],
-        toolRegistry: ToolRegistry,
-        toolSchemas: [ToolSchema],
-        inferenceOptions: InferenceOptions,
-        systemPrompt: String,
-        observer: (any AgentObserver)?,
-        tracing: TracingHelper?,
-        resultBuilder: AgentResult.Builder,
-        executionContext: AgentContext,
-        startTime: ContinuousClock.Instant,
-        session: (any Session)?,
-        enableStreaming: Bool,
-        structuredOutputRequest: StructuredOutputRequest?
-    ) async throws -> ToolLoopOutcome? {
-        guard configuration.foundationModelsExecution == .nativeSession else {
-            return nil
-        }
-
-        #if canImport(FoundationModels)
-        guard #available(macOS 26.0, iOS 26.0, visionOS 26.0, *) else {
-            return nil
-        }
-        #if os(tvOS) || os(watchOS)
-        return nil
-        #else
-        guard let fmProvider = provider as? FoundationModelsInferenceProvider else {
-            return nil
-        }
-
-        let options = optionsWithMembraneRuntimeSettings(inferenceOptions)
-        await observer?.onLLMStart(
-            context: nil,
-            agent: self,
-            systemPrompt: systemPrompt,
-            inputMessages: messages.map { message in
-                MemoryMessage(role: memoryRole(for: message.role), content: message.content)
-            }
-        )
-
-        let executingTools: [any FoundationModels.Tool]
-        if toolSchemas.isEmpty {
-            executingTools = []
-        } else {
-            let runtime = FoundationModelsNativeToolRuntime(
-                registry: toolRegistry,
-                agent: self,
-                context: executionContext,
-                observer: observer,
-                tracing: tracing,
-                resultBuilder: resultBuilder,
-                stopOnToolError: configuration.stopOnToolError
-            )
-            do {
-                executingTools = try FoundationModelsToolBridge.makeExecutingTools(
-                    from: toolSchemas,
-                    runtime: runtime
-                )
-            } catch {
-                throw AgentError.generationFailed(
-                    reason: "Failed to bridge Swarm tools to FoundationModels.Tool: \(error)"
-                )
-            }
-        }
-
-        let conversationID = session?.sessionId ?? "agent:\(configuration.name)"
-        let streamObserver = observer
-        let streamAgent: any AgentRuntime = self
-        let onOutputChunk: (@Sendable (String) async -> Void)?
-        if enableStreaming {
-            onOutputChunk = { chunk in
-                if let streamObserver {
-                    await streamObserver.onOutputToken(context: nil, agent: streamAgent, token: chunk)
-                }
-            }
-        } else {
-            onOutputChunk = nil
-        }
-        let native: FoundationModelsNativeTurnResult = try await executeWithinRemainingTimeout(startTime: startTime) {
-            try await fmProvider.respondUsingNativeSession(
-                messages: messages,
-                tools: executingTools,
-                toolSchemas: toolSchemas,
-                options: options,
-                conversationID: conversationID,
-                onOutputChunk: onOutputChunk
-            )
-        }
-
-        let finalResponse = try finalizeAssistantResponse(
-            content: native.content,
-            request: structuredOutputRequest,
-            provider: provider
-        )
-        await observer?.onLLMEnd(
-            context: nil,
-            agent: self,
-            response: finalResponse.content,
-            usage: nil
-        )
-
-        var transcriptMessages = native.transcriptMessages
-        if let structured = finalResponse.structuredOutput {
-            if let last = transcriptMessages.indices.last,
-               transcriptMessages[last].role == MemoryMessage.Role.assistant
-            {
-                transcriptMessages[last] = SwarmTranscriptCodec.encodeMessage(
-                    role: .assistant,
-                    content: finalResponse.content,
-                    structuredOutput: structured
-                )
-            }
-        }
-
-        return ToolLoopOutcome(
-            output: finalResponse.content,
-            structuredOutput: finalResponse.structuredOutput,
-            transcriptMessages: transcriptMessages
-        )
-        #endif
-        #else
-        return nil
-        #endif
-    }
-
-    private func memoryRole(for role: InferenceMessage.Role) -> MemoryMessage.Role {
-        switch role {
-        case .system: .system
-        case .user: .user
-        case .assistant: .assistant
-        case .tool: .tool
-        }
-    }
-
     /// Checks for cancellation and timeout conditions.
     private func checkCancellationAndTimeout(startTime: ContinuousClock.Instant) throws {
         // Use Task.checkCancellation() for reliable cancellation detection
@@ -1816,7 +1673,7 @@ public struct Agent: AgentRuntime, Sendable {
         }
     }
 
-    private func executeWithinRemainingTimeout<T: Sendable>(
+    func executeWithinRemainingTimeout<T: Sendable>(
         startTime: ContinuousClock.Instant,
         operation: @escaping @Sendable () async throws -> T
     ) async throws -> T {
@@ -2728,7 +2585,7 @@ public struct Agent: AgentRuntime, Sendable {
         )
     }
 
-    private func optionsWithMembraneRuntimeSettings(_ base: InferenceOptions) -> InferenceOptions {
+    func optionsWithMembraneRuntimeSettings(_ base: InferenceOptions) -> InferenceOptions {
         guard let membrane = AgentEnvironmentValues.current.membrane, membrane.isEnabled else {
             return base
         }
