+++ Tests/SwarmTests/Providers/FoundationModelsExecutionModeTests.swift
@@ -59,4 +59,40 @@ struct FoundationModelsExecutionModeTests {
         let recorded = await mockProvider.toolCallMessageCalls
         #expect(recorded.count == 2)
     }
+
+    @Test("Native session uses the envelope prompt when structured messages are nil")
+    func nativeSessionUsesEnvelopeWhenStructuredMessagesNil() {
+        let envelope = """
+        [Retrieved Context]
+        windowed-memory
+        [Current Conversation]
+        User: needle-user-input
+        """
+        let history = [
+            InferenceMessage.user("stale-turn-1"),
+            InferenceMessage.assistant("stale-turn-2"),
+            InferenceMessage.user("needle-user-input"),
+        ]
+
+        let messages = FoundationModelsNativePrompt.messages(
+            structuredMessages: nil,
+            envelopePrompt: envelope
+        )
+
+        #expect(messages == [.user(envelope)])
+        #expect(!messages.contains(where: { $0.content == history[0].content }))
+    }
+
+    @Test("Native session keeps structured messages when the envelope was not rewritten")
+    func nativeSessionKeepsStructuredMessages() {
+        let structured = [
+            InferenceMessage.user("hello"),
+            InferenceMessage.assistant("hi"),
+        ]
+        let messages = FoundationModelsNativePrompt.messages(
+            structuredMessages: structured,
+            envelopePrompt: "STUFFED SHOULD BE IGNORED"
+        )
+        #expect(messages == structured)
+    }
 }
